Extract EPNs into component with reusable ones
